fix a bug & introduce a new one
This commit fixes a bug where we weren't updating our state matrices
to reflect our initial hands, so we didn't actually know which cards
we had in hand.

It also revealed a new bug, now documented in `bug.md`, where we
erroneously pick up a card we don't want from the discard and then
immediately discard it.