when not `playVerbose`, print some output to stderr
namely, each game will produce a line of 0/1/x, where a number
indicates that player won a round, and an x indicates a round reseting
because the deck was exhausted, followed by -> 0/1, indicating the
winner of the game.

i'm running this with `make run | tee results.txt`, so i get the
stderr output in my terminal and the juicy stuff in a file.